Move NipsaFilter into search
