+++ h/nipsa/__init__.py
@@ -7,6 +7,3 @@ def includeme(config):
     # written into annotations on save.
     config.add_subscriber('h.nipsa.subscribers.transform_annotation',
                           'h.events.AnnotationTransformEvent')
-
-    # Register an additional filter with the API search module
-    config.add_search_filter('h.nipsa.search.Filter')
