+++ h/search/core.py
@@ -155,6 +155,7 @@ class Search(object):
         builder.append_filter(query.GroupFilter())
         builder.append_filter(query.GroupAuthFilter(request))
         builder.append_filter(query.UserFilter())
+        builder.append_filter(query.NipsaFilter(request))
         builder.append_matcher(query.AnyMatcher())
         builder.append_matcher(query.TagsMatcher())
         for factory in request.registry.get(FILTERS_KEY, []):
